requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1

# LangGraph and LangChain
langgraph==0.2.28
langchain==0.2.16
langchain-core>=0.2.39

# Database
psycopg2-binary==2.9.9
pgvector==0.2.5
redis==5.0.3
sqlalchemy==2.0.29

# Embeddings (self-hosted)
sentence-transformers==2.7.0
torch==2.3.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.20.0
playwright==1.44.0
lxml==5.2.1

# PDF parsing
pypdf==4.2.0
pdfplumber==0.11.0

# Data processing
pandas==2.2.2
numpy==1.26.4

# Utilities
python-dateutil==2.9.0
pytz==2024.1
schedule==1.2.1

# Logging
loguru==0.7.2
flask

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
